Fixing principal frame - works for regular tetrahedra

+++ OpenGLtests/Cylinder.h
@@ -14,16 +14,16 @@
 class Cylinder {
 
 public:
-	std::vector<GLfloat> vertices;
+	std::vector<GLdouble> vertices;
 
-	Cylinder(float radius, float height)  {
-		GLfloat TWO_PI = glm::pi<float>() * 2.0f;
-		GLfloat diffBetweenCircles = .01f;
+	Cylinder(double radius, double height)  {
+		GLdouble TWO_PI = glm::pi<double>() * 2.0f;
+		GLdouble diffBetweenCircles = .01f;
 		int numCircles = (int) height / diffBetweenCircles;
-		GLfloat radInc = glm::pi<float>() / 128.0f;
+		GLdouble radInc = glm::pi<double>() / 128.0f;
 
-		std::vector<GLfloat> circleXpts;
-		std::vector<GLfloat> circleZpts;
+		std::vector<GLdouble> circleXpts;
+		std::vector<GLdouble> circleZpts;
 		ShapeUtils::getCirclePoints(radius, circleXpts, circleZpts);
 		int circlePts = circleXpts.size();
 
@@ -31,15 +31,15 @@ public:
 	
 		for (int h = 0 - heightOffset; h < numCircles - heightOffset; h++){
 			for (int circ = 0; circ < circlePts; circ++){
-				GLfloat x1 = circleXpts[circ];
-				GLfloat y1 = diffBetweenCircles * h;
-				GLfloat z1 = circleZpts[circ];
+				GLdouble x1 = circleXpts[circ];
+				GLdouble y1 = diffBetweenCircles * h;
+				GLdouble z1 = circleZpts[circ];
 
-				GLfloat x2 = x1;
-				GLfloat y2 = y1 + diffBetweenCircles;
-				GLfloat z2 = z1;
+				GLdouble x2 = x1;
+				GLdouble y2 = y1 + diffBetweenCircles;
+				GLdouble z2 = z1;
 
-				GLfloat x3, z3;
+				GLdouble x3, z3;
 				if (circ == circlePts - 1){
 					x3 = circleXpts[0];
 					z3 = circleZpts[0];
@@ -48,24 +48,24 @@ public:
 					x3 = circleXpts[circ + 1];
 					z3 = circleZpts[circ + 1];
 				}
-				GLfloat y3 = y1;
+				GLdouble y3 = y1;
 
-				GLfloat x4 = x3;
-				GLfloat y4 = y2;
-				GLfloat z4 = z3;
+				GLdouble x4 = x3;
+				GLdouble y4 = y2;
+				GLdouble z4 = z3;
 
-				glm::vec3 p1(x1, y1, z1);
-				glm::vec3 p2(x2, y2, z2);
-				glm::vec3 p3(x3, y3, z3);
-				glm::vec3 p4(x4, y4, z4);
+				glm::dvec3 p1(x1, y1, z1);
+				glm::dvec3 p2(x2, y2, z2);
+				glm::dvec3 p3(x3, y3, z3);
+				glm::dvec3 p4(x4, y4, z4);
 
 				// Triangles look like:
 				///  p2 -- p4
 				///  |  \  |
 				///  p1 -- p3
 
-				glm::vec3 normal1 = ShapeUtils::getNormalOfTriangle(p1, p2, p3);
-				glm::vec3 normal2 = ShapeUtils::getNormalOfTriangle(p3, p2, p4);
+				glm::dvec3 normal1 = ShapeUtils::getNormalOfTriangle(p1, p2, p3);
+				glm::dvec3 normal2 = ShapeUtils::getNormalOfTriangle(p3, p2, p4);
 
 				ShapeUtils::addTriangleToVector(p1, p2, p3, normal1, vertices);
 				ShapeUtils::addTriangleToVector(p3, p2, p4, normal2, vertices);
@@ -73,15 +73,15 @@ public:
 		}
 	}
 
-	static TetrahedralMesh MeshCylinder(float radius, float height){
+	static TetrahedralMesh MeshCylinder(double radius, double height){
 		TetrahedralMesh mesh;
-		GLfloat TWO_PI = glm::pi<float>() * 2.0f;
-		GLfloat diffBetweenCircles = .01f;
+		GLdouble TWO_PI = glm::pi<double>() * 2.0f;
+		GLdouble diffBetweenCircles = .01f;
 		int numCircles = (int)height / diffBetweenCircles;
-		GLfloat radInc = glm::pi<float>() / 128.0f;
+		GLdouble radInc = glm::pi<double>() / 128.0f;
 
-		std::vector<GLfloat> circleXpts;
-		std::vector<GLfloat> circleZpts;
+		std::vector<GLdouble> circleXpts;
+		std::vector<GLdouble> circleZpts;
 		ShapeUtils::getCirclePoints(radius, circleXpts, circleZpts);
 		int circlePts = circleXpts.size();
 
@@ -89,15 +89,15 @@ public:
 
 		for (int h = 0 - heightOffset; h < numCircles - heightOffset; h++){
 			for (int circ = 0; circ < circlePts; circ++){
-				GLfloat x1 = circleXpts[circ];
-				GLfloat y1 = diffBetweenCircles * h;
-				GLfloat z1 = circleZpts[circ];
+				GLdouble x1 = circleXpts[circ];
+				GLdouble y1 = diffBetweenCircles * h;
+				GLdouble z1 = circleZpts[circ];
 
-				GLfloat x2 = x1;
-				GLfloat y2 = y1 + diffBetweenCircles;
-				GLfloat z2 = z1;
+				GLdouble x2 = x1;
+				GLdouble y2 = y1 + diffBetweenCircles;
+				GLdouble z2 = z1;
 
-				GLfloat x3, z3;
+				GLdouble x3, z3;
 				if (circ == circlePts - 1){
 					x3 = circleXpts[0];
 					z3 = circleZpts[0];
@@ -106,28 +106,28 @@ public:
 					x3 = circleXpts[circ + 1];
 					z3 = circleZpts[circ + 1];
 				}
-				GLfloat y3 = y1;
+				GLdouble y3 = y1;
 
-				GLfloat x4 = x3;
-				GLfloat y4 = y2;
-				GLfloat z4 = z3;
+				GLdouble x4 = x3;
+				GLdouble y4 = y2;
+				GLdouble z4 = z3;
 
-				glm::vec3 p1(x1, y1, z1);
-				glm::vec3 p2(x2, y2, z2);
-				glm::vec3 p3(x3, y3, z3);
-				glm::vec3 p4(x4, y4, z4);
+				glm::dvec3 p1(x1, y1, z1);
+				glm::dvec3 p2(x2, y2, z2);
+				glm::dvec3 p3(x3, y3, z3);
+				glm::dvec3 p4(x4, y4, z4);
 
 				//TETRAHEDRON:
-				glm::vec3 d1(0.0f, y1, 0.0f);
-				glm::vec3 d2(0.0f, y2, 0.0f);
+				glm::dvec3 d1(0.0f, y1, 0.0f);
+				glm::dvec3 d2(0.0f, y2, 0.0f);
 
 				// Triangles look like:
 				///  p2 -- p4
 				///  |  \  |
 				///  p1 -- p3
 
-				glm::vec3 normal1 = ShapeUtils::getNormalOfTriangle(p1, p2, p3);
-				glm::vec3 normal2 = ShapeUtils::getNormalOfTriangle(p3, p2, p4);
+				glm::dvec3 normal1 = ShapeUtils::getNormalOfTriangle(p1, p2, p3);
+				glm::dvec3 normal2 = ShapeUtils::getNormalOfTriangle(p3, p2, p4);
 
 				//ShapeUtils::addTriangleToVector(p1, p2, p3, normal1, vertices);
 				//ShapeUtils::addTriangleToVector(p3, p2, p4, normal2, vertices);
